spark/11-Distributed-Concurrency/requirements.txt
# Core Dependencies
pyspark==3.4.0
pandas==2.1.4
pyarrow==14.0.2

# Ray for Distributed Computing
ray[default]==2.9.0
ray[tune]==2.9.0
ray[train]==2.9.0

# Celery for Task Queue
celery==5.3.4
redis==5.0.1

# Kafka
kafka-python==2.0.2

# MLflow for Experiment Tracking
mlflow==2.9.2

# Zookeeper
kazoo==2.9.0

# Monitoring
prometheus-client==0.19.0

# Web Framework
fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic==2.5.3

# Utilities
python-dotenv==1.0.0
psycopg2-binary==2.9.9
